backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# AI/ML
torch==2.2.1
torchvision==0.17.1
ultralytics==8.1.18
openai==1.12.0
pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.4

# Storage
boto3==1.34.44
minio==7.2.4

# Utils
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
requests==2.31.0
aiofiles==23.2.1

# Logging & Monitoring
structlog==24.1.0
prometheus-client==0.19.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.27.0
